Added the IReducible<T> interface...
..and implemented it for NOPList, array and FingerTree. This
is one of the most powerful and elegant functional constructs
I have added so far.

Also added basic tests for finger trees, and it bloody well
works. Pretty amazing for a first time run. The finger tree
data structure is so well typed that it is hard to add bugs
to the implementation :-)