The first version for reading and writing S-expressions is now done.
The reading utilizes the parsing library, which seems to work like
a charm. It uses context information give nice error messages. In
theory the current implementation could serve as the reader part
of REPL functionality.

Need to fine tune the parsing stuff still a bit, but in general the
decision to write that seemed to be correct. The only drawback
is debugging it in Visual Studio which, because of the monadic
nature of parsing, is simply confusing. Luckily the only problems
I had to debug were because of the static field initializers, which
as usual gave few surprises.