Played around with collection interface IStream and ISequence.
Trying to find a balance between operations defined in concrete
collection classes in generic interfaces and generic extension
methods. The rule of thumb has been that, if an operation is
inefficient in some collection class when defined generically,
then there should be a specific implementation that is optimized
for that collection type. The operation then becomes part of the
ISequence interface that contains the most popular list transformation
functions.